+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - PR #435: Exposing cumlhandle in cython + developer guide
 - PR #455: Remove default-stream arguement across ml-prims and cuML
 - PR #375: cuml cpp shared library renamed to libcuml++.so
+- PR #460: Random Forest & Decision Trees (Single-GPU, Classification
 - PR #491: added doxygen build target for ml-prims
 - PR #520: Add local build script to mimic gpuCI
 - PR #503: Added column-wise matrix sort primitive